@@ -2,8 +2,8 @@ from setuptools import find_namespace_packages, setup
 
 setup(
     name="NEMO",
-    version="6.0.2",
-    python_requires=">=3.8, <4",
+    version="6.1.0.dev",
+    python_requires=">=3.9, <4",
     packages=find_namespace_packages(exclude=["resources", "resources.*", "build", "build.*"]),
     include_package_data=True,
     url="https://github.com/usnistgov/NEMO",
@@ -20,28 +20,28 @@ setup(
         "License :: Public Domain",
         "Natural Language :: English",
         "Operating System :: OS Independent",
-        "Framework :: Django :: 3.2",
-        "Programming Language :: Python :: 3.8",
+        "Framework :: Django :: 4.2",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
+        "Programming Language :: Python :: 3.13",
     ],
     install_requires=[
         "cryptography==42.0.8",
-        "Django==4.2.11",
+        "Django==4.2.15",
         "django-auditlog==3.0.0",
         "django-filter==23.5",
         "django-mptt==0.14.0",
-        "djangorestframework==3.15.1",
+        "djangorestframework==3.15.2",
         "drf-excel==2.4.0",
         "drf-flex-fields==1.0.2",
         "ldap3==2.9.1",
-        "Pillow==10.3.0",
+        "Pillow==11.0.0",
         "pymodbus==3.3.2",
         "python-dateutil==2.9.0",
-        "pytz==2024.1",
         "requests==2.32.3",
+        "packaging==24.0.0",
     ],
     extras_require={"dev-tools": ["pre-commit", "djlint", "black"]},
     entry_points={
