@@ -3,6 +3,7 @@ from http import HTTPStatus
 
 from django.conf import settings
 from django.contrib.auth.decorators import login_required, permission_required
+from django.core.exceptions import ValidationError
 from django.db.models import Q
 from django.shortcuts import redirect, render
 from django.utils import timezone
@@ -92,6 +93,13 @@ def do_enable_tool(request, tool_id):
     except RequiredUnansweredQuestionsException as e:
         dictionary = {"message": str(e), "delay": 10}
         return render(request, "kiosk/acknowledgement.html", dictionary)
+
+    # Validate usage event
+    try:
+        new_usage_event.full_clean()
+    except ValidationError as e:
+        return render(request, "kiosk/acknowledgement.html", {"message": str(e)})
+
     new_usage_event.save()
 
     # Remove wait list entry if it exists
@@ -100,11 +108,10 @@ def do_enable_tool(request, tool_id):
         wait_list_entry.update(deleted=True, date_exited=timezone.now())
 
     try:
-        dynamic_form.charge_for_consumables(new_usage_event, new_usage_event.pre_run_data, request)
+        dynamic_form.process_run_data(new_usage_event, new_usage_event.pre_run_data, request)
     except Exception as e:
         dictionary = {"message": str(e), "delay": 10}
         return render(request, "kiosk/acknowledgement.html", dictionary)
-    dynamic_form.update_tool_counters(new_usage_event.pre_run_data, tool.id)
 
     dictionary = {"message": "You can now use the {}".format(tool), "badge_number": customer.badge_number}
     return render(request, "kiosk/acknowledgement.html", dictionary)
@@ -151,8 +158,8 @@ def do_disable_tool(request, tool_id):
     try:
         current_usage_event.run_data = dynamic_form.extract(request)
     except RequiredUnansweredQuestionsException as e:
-        if customer.is_staff and customer != current_usage_event.operator and current_usage_event.user != customer:
-            # if a staff is forcing somebody off the tool and there are required questions, send an email and proceed
+        if customer != current_usage_event.operator and current_usage_event.user != customer:
+            # if someone else is forcing somebody off the tool and there are required questions, send an email and proceed
             current_usage_event.run_data = e.run_data
             email_managers_required_questions_disable_tool(current_usage_event.operator, customer, tool, e.questions)
         else:
@@ -160,11 +167,10 @@ def do_disable_tool(request, tool_id):
             return render(request, "kiosk/acknowledgement.html", dictionary)
 
     try:
-        dynamic_form.charge_for_consumables(current_usage_event, current_usage_event.run_data, request)
+        dynamic_form.process_run_data(current_usage_event, current_usage_event.run_data, request)
     except Exception as e:
         dictionary = {"message": str(e), "delay": 10}
         return render(request, "kiosk/acknowledgement.html", dictionary)
-    dynamic_form.update_tool_counters(current_usage_event.run_data, tool.id)
 
     current_usage_event.save()
     dictionary = {
@@ -255,11 +261,14 @@ def reserve_tool(request):
 
     """ Create a reservation for a user. """
     try:
-        date = parse_date(request.POST["date"])
-        start = localize(datetime.combine(date, parse_time(request.POST["start"])))
-        end = localize(datetime.combine(date, parse_time(request.POST["end"])))
+        start_date = parse_date(request.POST["start_date"])
+        end_date = parse_date(request.POST["end_date"])
+        start = localize(datetime.combine(start_date, parse_time(request.POST["start"])))
+        end = localize(datetime.combine(end_date, parse_time(request.POST["end"])))
     except:
-        dictionary["message"] = "Please enter a valid date, start time, and end time for the reservation."
+        dictionary["message"] = (
+            "Please enter a valid start date, start time, end date and end time for the reservation."
+        )
         return render(request, "kiosk/error.html", dictionary)
     # Create the new reservation:
     reservation = Reservation()
