@@ -1,4 +1,4 @@
-from copy import deepcopy
+import json
 
 from django import forms
 from django.contrib import admin, messages
@@ -136,6 +136,14 @@ class SensorAdminForm(forms.ModelForm):
         model = Sensor
         fields = "__all__"
 
+    def clean_extra_args(self):
+        extra_args = self.cleaned_data["extra_args"]
+        try:
+            return json.dumps(json.loads(extra_args), indent=4)
+        except:
+            pass
+        return extra_args
+
     def clean(self):
         if any(self.errors):
             return
@@ -211,8 +219,8 @@ class SensorAdmin(admin.ModelAdmin):
         "read_address",
         "number_of_values",
         "get_read_frequency",
-        "get_last_read",
-        "get_last_read_at",
+        "last_read",
+        "last_value",
     )
     list_filter = (
         "visible",
@@ -222,20 +230,11 @@ class SensorAdmin(admin.ModelAdmin):
     )
     actions = [duplicate_sensor_configuration, read_selected_sensors, hide_selected_sensors, show_selected_sensors]
     autocomplete_fields = ["sensor_card", "interlock_card"]
+    readonly_fields = ["last_read", "last_value"]
 
     @display(boolean=True, ordering="sensor_card__enabled", description="Card Enabled")
     def get_card_enabled(self, obj: Sensor):
         return obj.card.enabled
-
-    @display(description="Last read")
-    def get_last_read(self, obj: Sensor):
-        last_data_point = obj.last_data_point()
-        return last_data_point.value if last_data_point else ""
-
-    @display(description="Last read at")
-    def get_last_read_at(self, obj: Sensor):
-        last_data_point = obj.last_data_point()
-        return last_data_point.created_date if last_data_point else ""
 
     @display(ordering="read_frequency", description="Read frequency")
     def get_read_frequency(self, obj: Sensor):
