@@ -1,5 +1,11 @@
 from NEMO.models import Area, Notification, PhysicalAccessLevel, Tool, User
-from NEMO.utilities import date_input_js_format, datetime_input_js_format, time_input_js_format
+from NEMO.utilities import (
+    date_input_js_format,
+    datetime_input_js_format,
+    time_input_js_format,
+    pickadate_date_format,
+    pickadate_time_format,
+)
 from NEMO.views.customization import CustomizationBase
 from NEMO.views.notifications import get_notification_counts
 
@@ -94,6 +100,8 @@ def base_context(request):
         "time_input_js_format": time_input_js_format,
         "date_input_js_format": date_input_js_format,
         "datetime_input_js_format": datetime_input_js_format,
+        "pickadate_date_format": pickadate_date_format,
+        "pickadate_time_format": pickadate_time_format,
         "no_header": request.session.get("no_header", False),
         "safety_menu_item": customization_values.get("safety_main_menu") == "enabled",
         "calendar_page_title": customization_values.get("calendar_page_title"),
