@@ -16,7 +16,11 @@ from django.urls.resolvers import RegexPattern
 from NEMO.models import User
 from NEMO.tests.test_utilities import login_as, login_as_staff, login_as_user, login_as_user_with_permissions
 from NEMO.utilities import get_full_url
-from NEMO.views.customization import ApplicationCustomization, EmailsCustomization, UserRequestsCustomization
+from NEMO.views.customization import (
+    AdjustmentRequestsCustomization,
+    ApplicationCustomization,
+    EmailsCustomization,
+)
 
 url_test_logger = getLogger(__name__)
 
@@ -137,6 +141,7 @@ url_kwargs_get_post = {
     "knowledge_base_categories": {"kwargs": {"kind": "user"}},
     "knowledge_base_all_in_one": {"kwargs": {"kind": "user"}},
     "view_user": {"login_id": 1},
+    "enable_tool": {"login_id": 1, "kwargs": {"tool_id": 3, "user_id": 1, "project_id": 1, "staff_charge": "false"}},
 }
 
 urls_to_skip = [
@@ -157,6 +162,7 @@ urls_to_skip = [
     "new_reservation",
     "media_view",
     "media_list_view",
+    "api_media",
     "remove_document_from_project",
     "enter_wait_list_from_kiosk",
     "exit_wait_list_from_kiosk",
@@ -174,7 +180,7 @@ class URLsTestCase(TestCase):
         EmailsCustomization.set("user_office_email_address", "email@example.org")
         EmailsCustomization.set("safety_email_address", "email@example.org")
         EmailsCustomization.set("abuse_email_address", "email@example.org")
-        UserRequestsCustomization.set("adjustment_requests_enabled", "enabled")
+        AdjustmentRequestsCustomization.set("adjustment_requests_enabled", "enabled")
 
     def test_get_full_url(self):
         request = RequestFactory().get("/")
@@ -221,7 +227,7 @@ class URLsTestCase(TestCase):
             "event_feed",
             {
                 "get": {
-                    "event_type": f"{facility_name.lower()} usage",
+                    "event_type": f"{facility_name.lower()} use",
                     "start": start.strftime("%Y-%m-%d"),
                     "end": end_one_day.strftime("%Y-%m-%d"),
                     "item_type": "tool",
@@ -234,7 +240,7 @@ class URLsTestCase(TestCase):
             "event_feed",
             {
                 "get": {
-                    "event_type": f"{facility_name.lower()} usage",
+                    "event_type": f"{facility_name.lower()} use",
                     "start": start.strftime("%Y-%m-%d"),
                     "end": end_one_day.strftime("%Y-%m-%d"),
                     "personal_schedule": "yes",
@@ -246,7 +252,7 @@ class URLsTestCase(TestCase):
             "event_feed",
             {
                 "get": {
-                    "event_type": f"{facility_name.lower()} usage",
+                    "event_type": f"{facility_name.lower()} use",
                     "start": start.strftime("%Y-%m-%d"),
                     "end": end_one_day.strftime("%Y-%m-%d"),
                     "all_tools": "yes",
@@ -258,7 +264,7 @@ class URLsTestCase(TestCase):
             "event_feed",
             {
                 "get": {
-                    "event_type": f"{facility_name.lower()} usage",
+                    "event_type": f"{facility_name.lower()} use",
                     "start": start.strftime("%Y-%m-%d"),
                     "end": end_one_day.strftime("%Y-%m-%d"),
                     "all_areas": "yes",
@@ -270,7 +276,7 @@ class URLsTestCase(TestCase):
             "event_feed",
             {
                 "get": {
-                    "event_type": f"{facility_name.lower()} usage",
+                    "event_type": f"{facility_name.lower()} use",
                     "start": start.strftime("%Y-%m-%d"),
                     "end": end_one_day.strftime("%Y-%m-%d"),
                     "all_areastools": "yes",
