@@ -1,14 +1,2 @@
-# Maximum length for additional information of reservations
-ADDITIONAL_INFORMATION_MAXIMUM_LENGTH = 3000
-
-# Maximum length for feedback emails
-FEEDBACK_MAXIMUM_LENGTH = 5000
-
-# Maximum length for Char fields to work with every db type
-CHAR_FIELD_MAXIMUM_LENGTH = 255
-
-# Name of the parameter to indicate which view to redirect to
-NEXT_PARAMETER_NAME = "next"
-
-# Name of the media folder under which only staff can see files
-MEDIA_PROTECTED = "protected"
+# Kept for backward compatibility. remove in future version of NEMO
+from NEMO.constants import *
