@@ -10,7 +10,7 @@ from django.utils import timezone
 from django.views.decorators.http import require_GET
 
 from NEMO.models import Alert, LandingPageChoice, Reservation, Resource, UsageEvent, User
-from NEMO.views.alerts import delete_expired_alerts
+from NEMO.views.alerts import mark_alerts_as_expired
 from NEMO.views.area_access import able_to_self_log_in_to_area, able_to_self_log_out_of_area
 from NEMO.views.notifications import delete_expired_notifications
 
@@ -19,7 +19,7 @@ from NEMO.views.notifications import delete_expired_notifications
 @require_GET
 def landing(request):
     user: User = request.user
-    delete_expired_alerts()
+    mark_alerts_as_expired()
     delete_expired_notifications()
     usage_events = UsageEvent.objects.filter(operator=user.id, end=None).prefetch_related("tool", "project")
     tools_in_use = [u.tool.tool_or_parent_id() for u in usage_events]
@@ -63,6 +63,7 @@ def landing(request):
         "landing_page_choices": landing_page_choices,
         "self_log_in": able_to_self_log_in_to_area(request.user),
         "self_log_out": able_to_self_log_out_of_area(request.user),
+        "script_name": settings.FORCE_SCRIPT_NAME,
     }
     return render(request, "landing.html", dictionary)
 
