@@ -3,7 +3,7 @@ from django.shortcuts import render
 from django.views.decorators.http import require_GET
 
 from NEMO.models import Area, PhysicalAccessLevel, User
-from NEMO.views.customization import UserRequestsCustomization
+from NEMO.views.customization import AdjustmentRequestsCustomization, UserRequestsCustomization
 
 
 @login_required
@@ -18,7 +18,7 @@ def user_requests(request, tab: str = None):
     buddy_requests_title = UserRequestsCustomization.get("buddy_requests_title")
     staff_assistance_requests_title = UserRequestsCustomization.get("staff_assistance_requests_title")
     access_requests_title = UserRequestsCustomization.get("access_requests_title")
-    adjustment_requests_title = UserRequestsCustomization.get("adjustment_requests_title")
+    adjustment_requests_title = AdjustmentRequestsCustomization.get("adjustment_requests_title")
     dictionary = {
         "tab": active_tab,
         "buddy_requests_title": buddy_requests_title,
